requirements.txt
openai

guardrails-ai

python-dotenv

lxml

jsonschema